HBP-87. Move examples from tvb_nest at top-level under examples/tvb_nest folder.
