+++ .gitignore
@@ -109,7 +109,7 @@ venv.bak/
 #tvb-nest outputs
 outputs/figs
 outputs/logs
-/tvb_multiscale/tvb_nest/examples/outputs/
-/tvb_multiscale/tvb_nest/examples/notebooks/reduced_wong_wang_exc_io_inh_i/ConceptGraph.png
-/tvb_multiscale/tvb_nest/examples/notebooks/reduced_wong_wang_exc_io_inh_i/documented_notebook_example.ipynb
-/tvb_multiscale/tvb_nest/examples/notebooks/reduced_wong_wang_exc_io_inh_i/simulator_loop.png
+/tvb_multiscale/examples/tvb_nest/outputs/
+/tvb_multiscale/examples/tvb_nest/notebooks/reduced_wong_wang_exc_io_inh_i/ConceptGraph.png
+/tvb_multiscale/examples/tvb_nest/notebooks/reduced_wong_wang_exc_io_inh_i/documented_notebook_example.ipynb
+/tvb_multiscale/examples/tvb_nest/notebooks/reduced_wong_wang_exc_io_inh_i/simulator_loop.png
