+++ docker/hpc/example_script.sh
@@ -9,9 +9,9 @@
 export DOCKER_HOME=/home/docker
 export DOCKER_TVB_ROOT=$DOCKER_HOME/packages/tvb-root
 export DOCKER_TVB_MULTISCALE=$DOCKER_HOME/packages/tvb-multiscale
-export DOCKER_WORKDIR=$DOCKER_TVB_MULTISCALE/tvb_nest/examples
+export DOCKER_WORKDIR=$DOCKER_TVB_MULTISCALE/tvb_multiscale/examples/tvb_nest
 export DOCKER_PYTHON=/home/docker/env/neurosci/bin/python
-export DOCKER_RUN_FILE=/home/docker/packages/tvb-multiscale/tvb_nest/examples/example.py
+export DOCKER_RUN_FILE=/home/docker/packages/tvb-multiscale/tvb_multiscale/examples/tvb_nest/example.py
 export CMD=$DOCKER_PYTHON' '$DOCKER_RUN_FILE
 
 export NODES=mc  # or gpu, nothing for BIH cluster
