+++ docker/launch_example.py
@@ -9,7 +9,7 @@ mpl.use('Agg')
 import os
 import numpy as np
 from tvb_multiscale.tvb_nest.config import Config
-from tvb_multiscale.tvb_nest.examples.example import main_example
+from tvb_multiscale.examples.tvb_nest.example import main_example
 from tvb_multiscale.tvb_nest.nest_models.builders.models.wilson_cowan import WilsonCowanBuilder
 from tvb_multiscale.tvb_nest.interfaces.builders.models.wilson_cowan \
     import WilsonCowanBuilder as InterfaceWilsonCowanBuilder
