+++ tvb_multiscale/tests/tvb_nest/test_models.py
@@ -6,7 +6,7 @@ TvbProfile.set_profile(TvbProfile.LIBRARY_PROFILE)
 import matplotlib as mpl
 mpl.use('Agg')
 
-from tvb_multiscale.tvb_nest.examples.example import main_example, results_path_fun
+from tvb_multiscale.examples.tvb_nest.example import main_example, results_path_fun
 from tvb_multiscale.tvb_nest.nest_models.builders.models.basal_ganglia_izhikevich import BasalGangliaIzhikevichBuilder
 from tvb_multiscale.tvb_nest.nest_models.builders.models.ww_deco import WWDeco2013Builder, WWDeco2014Builder
 from tvb_multiscale.tvb_nest.nest_models.builders.models.wilson_cowan import WilsonCowanBuilder, WilsonCowanMultisynapseBuilder
